+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Changed
 
+- Add "macOS" to System Information → Software → macOS → Code Name
+
 ### Fixed
 
 - Disable app updates in recoveryOS as Sparkle does not work there